+++ src/main/scala/arihunta/advent2019/Day05.scala
@@ -10,21 +10,21 @@ import java.util.concurrent.LinkedBlockingQueue
 object Day05 {
 
     def _01(): Int = {
-        val program = Source.fromResource("05").mkString.trim().split(",").map(it => it.toInt)
-        val input: BlockingQueue[Int] = new LinkedBlockingQueue
+        val program = Source.fromResource("05").mkString.trim().split(",").map(it => BigInt(it))
+        val input: BlockingQueue[BigInt] = new LinkedBlockingQueue
         input.add(1)
         val computer = new IntCodeComputer(program, input)
         computer.run()
-        computer.output.toArray()(computer.output.size() - 1).asInstanceOf[Int]
+        computer.output.toArray()(computer.output.size() - 1).asInstanceOf[BigInt].toInt
     }
 
     def _02(): Int = {
-        val program = Source.fromResource("05").mkString.trim().split(",").map(it => it.toInt)
-        val input: BlockingQueue[Int] = new LinkedBlockingQueue
+        val program = Source.fromResource("05").mkString.trim().split(",").map(it => BigInt(it))
+        val input: BlockingQueue[BigInt] = new LinkedBlockingQueue
         input.add(5)
         val computer = new IntCodeComputer(program, input)
         computer.run()
-        computer.output.toArray()(computer.output.size() - 1).asInstanceOf[Int]
+        computer.output.toArray()(computer.output.size() - 1).asInstanceOf[BigInt].toInt
     }
 
 }
