+++ src/main/scala/arihunta/advent2019/Day07.scala
@@ -10,7 +10,7 @@ object Day07 {
 
     def _01() = {
 
-        val program = Source.fromResource("07").mkString.trim().split(",").map(it => it.toInt)
+        val program = Source.fromResource("07").mkString.trim().split(",").map(it => BigInt(it))
 
         generateCombinations(Array(0, 1, 2, 3, 4))
             .map(it => runTest(program, it))
@@ -20,7 +20,7 @@ object Day07 {
 
     def _02() = {
 
-        val program = Source.fromResource("07").mkString.trim().split(",").map(it => it.toInt)
+        val program = Source.fromResource("07").mkString.trim().split(",").map(it => BigInt(it))
 
         generateCombinations(Array(5, 6, 7, 8, 9))
             .map(it => runTestFeedbackLoop(program, it))
@@ -28,7 +28,7 @@ object Day07 {
 
     }
 
-    def generateCombinations(input: Array[Int]): Array[Array[Int]] = {
+    def generateCombinations(input: Array[BigInt]): Array[Array[BigInt]] = {
         if (input.size == 1)
             Array(input)
         else {
@@ -39,14 +39,14 @@ object Day07 {
         }
     }
 
-    private def runTest(program: Array[Int], inputs: Array[Int]): Int = {
+    private def runTest(program: Array[BigInt], inputs: Array[BigInt]): Int = {
 
-        val _in01: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(0)))
-        val _out01_in02: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(1)))
-        val _out02_in03: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(2)))
-        val _out03_in04: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(3)))
-        val _out04_in05: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(4)))
-        val _out05: BlockingQueue[Int] = new LinkedBlockingQueue()
+        val _in01: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(0)))
+        val _out01_in02: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(1)))
+        val _out02_in03: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(2)))
+        val _out03_in04: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(3)))
+        val _out04_in05: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(4)))
+        val _out05: BlockingQueue[BigInt] = new LinkedBlockingQueue()
 
         _in01.put(0)
 
@@ -62,17 +62,17 @@ object Day07 {
         amp04.runOnThread()
         amp05.runOnThread()
 
-        amp05.output.take()
+        amp05.output.take().toInt
 
     }
 
-    private def runTestFeedbackLoop(program: Array[Int], inputs: Array[Int]): Int = {
+    private def runTestFeedbackLoop(program: Array[BigInt], inputs: Array[BigInt]): Int = {
 
-        val _out05_in01: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(0)))
-        val _out01_in02: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(1)))
-        val _out02_in03: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(2)))
-        val _out03_in04: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(3)))
-        val _out04_in05: BlockingQueue[Int] = new LinkedBlockingQueue(Arrays.asList(inputs(4)))
+        val _out05_in01: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(0)))
+        val _out01_in02: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(1)))
+        val _out02_in03: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(2)))
+        val _out03_in04: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(3)))
+        val _out04_in05: BlockingQueue[BigInt] = new LinkedBlockingQueue(Arrays.asList(inputs(4)))
 
         _out05_in01.put(0)
 
@@ -96,7 +96,7 @@ object Day07 {
                 thread05,
         ).foreach(it => it.join())
 
-        amp05.output.remove()
+        amp05.output.remove().toInt
 
     }
